+++ mobile/android/android-components/components/browser/state/src/main/java/mozilla/components/browser/state/action/BrowserAction.kt
@@ -153,6 +153,22 @@ sealed class TabListAction : BrowserAction() {
     data class AddMultipleTabsAction(val tabs: List<TabSessionState>) : TabListAction()
 
     /**
+     * Moves a set of tabIDs into a new position (maintaining internal ordering)
+     *
+     * @property tabIds The IDs of the tabs to move.
+     * @property targetTabId A tab that the moved tabs will be moved next to
+     * @property placeAfter True if the moved tabs should be placed after the target,
+     * False for placing before the target. Irrelevant if the target is one of the tabs being moved,
+     * since then the whole list is moved to where the target was. Ordering of the moved tabs
+     * relative to each other is preserved.
+     */
+    data class MoveTabsAction(
+        val tabIds: List<String>,
+        val targetTabId: String,
+        val placeAfter: Boolean,
+    ) : TabListAction()
+
+    /**
      * Marks the [TabSessionState] with the given [tabId] as selected tab.
      *
      * @property tabId the ID of the tab to select.