+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/TabsUseCases.kt
@@ -478,6 +478,34 @@ class TabsUseCases(
         }
     }
 
+    /**
+     * Use case for moving a collection of tabs.
+     */
+    class MoveTabsUseCase(
+        private val store: BrowserStore
+    ) {
+        /**
+         * Moves the tabs of [tabIds] next to [targetTabId], before/after based on [placeAfter]
+         *
+         * @property tabIds The IDs of the tabs to move.
+         * @property targetTabId A tab that the moved tabs will be moved next to
+         * @property placeAfter True if the moved tabs should be placed after the target,
+         * False for placing before the target. Irrelevant if the target is one of the tabs being moved,
+         * since then the whole list is moved to where the target was. Ordering of the moved tabs
+         * relative to each other is preserved.
+         */
+        operator fun invoke(
+            tabIds: List<String>,
+            targetTabId: String,
+            placeAfter: Boolean
+        ) {
+            store.dispatch(
+                TabListAction.MoveTabsAction(
+                    tabIds, targetTabId, placeAfter
+                )
+            )
+        }
+    }
     val selectTab: SelectTabUseCase by lazy { DefaultSelectTabUseCase(store) }
     val removeTab: RemoveTabUseCase by lazy { DefaultRemoveTabUseCase(store) }
     val addTab: AddNewTabUseCase by lazy { AddNewTabUseCase(store) }
@@ -492,4 +520,5 @@ class TabsUseCases(
     val restore: RestoreUseCase by lazy { RestoreUseCase(store, selectTab) }
     val selectOrAddTab: SelectOrAddUseCase by lazy { SelectOrAddUseCase(store) }
     val duplicateTab: DuplicateTabUseCase by lazy { DuplicateTabUseCase(store) }
+    val moveTabs: MoveTabsUseCase by lazy { MoveTabsUseCase(store) }
 }